+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [fj-doc-val-p7m] P7MPemValidator and P7MRawValidator
 
+### Changed
+
+- [fj-doc-val-p7m] Optimized P7MValidator (for previous behaviour use P7MLegacyValidator)
+
 ## [8.7.6] - 2024-09-02
 
 ### Fixed